pyproject.toml
[tool.poetry]
name = "4desa-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
